Merge 'Test' and 'Publish' workflows
The 'workflow_run' event triggers a workflow run by given the last
commit on the default branch as reference, which is incompatible with
the environment protection rules in place. Per the environment
protection rules, only an event occurring on the 'master' branch can
trigger the 'Publish' workflow whereas the default branch is 'develop'.

This commit attempts to fix the issue by merging the two workflows into
a single one.